requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
chromadb
sentence-transformers
pypdf
pytube
python-dotenv
